requirements.txt
streamlit>=1.31,<2.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0
xlsxwriter>=3.1,<4.0
requests>=2.31,<3.0
sqlalchemy>=2.0,<3.0
alembic>=1.12,<2.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
